results/giant: move giant REST API functions to the library
